Crate axtask

ArceOS task management module.

This module provides primitives for task management, including task creation, scheduling, sleeping, termination, etc. The scheduler algorithm is configurable by cargo features.

Cargo Features

multitask: Enable multi-task support. If it’s enabled, complex task management and scheduling is used, as well as more task-related APIs. Otherwise, only a few APIs with naive implementation is available.

irq: Interrupts are enabled. If this feature is enabled, timer-based APIs can be used, such as sleep, sleep_until, and WaitQueue::wait_timeout.

preempt: Enable preemptive scheduling.

sched_fifo: Use the FIFO cooperative scheduler. It also enables the multitask feature if it is enabled. This feature is enabled by default, and it can be overriden by other scheduler features.

sched_rr: Use the Round-robin preemptive scheduler. It also enables the multitask and preempt features if it is enabled.

sched_cfs: Use the Completely Fair Scheduler. It also enables the the multitask and preempt features if it is enabled.

Macros

declare_wait

A queue to store sleeping tasks.

Structs

CurrentTaskmultitask

A wrapper of AxTaskRef as the current task.

Processor

Processor on coresponding core

SchedStatus

The status of the scheduler

TaskIdmultitask

A unique identifier for a thread.

WaitQueuemultitask

A queue to store tasks that are waiting for some conditions.

WaitTaskList

A simple FIFO wait task list

WaitTaskNode

A task wrapper.

Enums

SchedPolicy

The policy of the scheduler

TaskState

The possible states of a task.

Constants

MAX_RT_PRIO

Functions

currentmultitask

Gets the current task.

current_check_preempt_pendingmultitask

Checks if the current task should be preempted. This api called after handle irq,it may be on a disable_preempt ctx

current_may_uninitmultitask

Gets the current task, or returns None if the current task is not initialized.

current_processor

Get current processor pointer

dump_curr_backtracemultitask

Dump the current task backtrace.

dump_task_backtracemultitask

Dump the given task backtrace.

exitmultitask

Exits the current task.

init_schedulermultitask

Initializes the task scheduler (for the primary CPU).

init_scheduler_secondarymultitask

Initializes the task scheduler for secondary CPUs.

joinmultitask

Current task is going to sleep, it will be woken up when the given task exits.

new_taskmultitask

Create a new task.

on_timer_tickmultitask and irq

Handles periodic timer ticks for the task manager.

run_idlemultitask

The idle task routine.

schedule

Schedule the next task to run.

schedule_timeoutirq

Sleep the current task until the deadline.

set_prioritymultitask

Set the priority for current task.

sleep

Current task is going to sleep for the given duration.

sleep_until

Current task is going to sleep, it will be woken up at the given deadline.

spawnmultitask

Spawns a new task with the default parameters.

spawn_rawmultitask

Spawns a new task with the given parameters.

vfork_suspendmultitask

Current task is going to sleep. It will be woken up when the given task does exec syscall or exit.

wake_vfork_processmultitask

To wake up the task that is blocked because vfork out of current task

wakeup_taskmultitask

wake up task

yield_now

Current task gives up the CPU time voluntarily, and switches to another ready task.

Type Aliases

AxTaskRefmultitask

The reference type of a task.