axtask

Trait TaskExtMut

pub trait TaskExtMut<T: Sized> {
    // Required method
    fn task_ext_mut(&mut self) -> &mut T;
}

Available on crate feature multitask only.

A trait to convert TaskInner::task_ext_ptr to the mutable reference of the concrete type.

Required Methods

fn task_ext_mut(&mut self) -> &mut T

Get a mutable reference to the task extended data.